Handle navbar highlighting through javascript
There are two different ways to handle navbar highlighting - and I'm
not certain which is better right now. This method uses javascript
and the `pageId` variable set in the YAML for each page. This keeps
the template rather clean by just allowing me to add an attribute
`id="navbar-PAGENAME"` to each navbar element. However, it means that
the navbar lights up a split second later, which is a bit off-putting.

Alternatively this logic can be put right into the template. It will
make the template messier, but it will make it highlight the appropriate
pages right away.

I also added in a `.ruby-version` file that makes stuff work with
middleman on my mac.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>