Initial BibTeX parsing
This adds a macro that can be used to parse out BibTeX files and insert
them into a web page. The general usage is something like this:

    <ul class="bibliography bibliography-conference">
       <%= bibtex_list("publications/papers.bib", "mypapers:journal", "J") %>
    </ul>

First parameter is the name of the BibTeX file relative to the root
of the source tree. Second parameter is the keyword to filter on. I use
keywords such as `mypapers:journal`, `mypapers:workshop`, etc to segment
out my papers by category. The third paramter is the prefix that should
be applied to the counter when counting down through the items.

This isn't 100% perfect yet. There are a few issues to address:

* Pedants will get annoyed because the style isn't fully compliant with
citeproc stuff. That may get updated in the future.
* Sorting isn't fully awesome yet. This is because the items are sorted
by date, but only out to the year. I need to find a better way to
parse out the dates so they're listed in proper order.
* This is also a bit of an arbitrary hacky process. Perhaps using
citeproc will fix some of those problems. I just need to figure out how
to have citeproc create the semantic elements and insert they links.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>