Initial commit of new Wintersmith version
This still lacks a lot of what the older middleman version had. In
particular, it doesn't have the fancy automatic image resizing for the
affiliations on the home page, bibliography support, or even my resume,
but it's a bit of a start of something that works on Windows and Mac,
which is important as I start to distance personal from professional
computing resources.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>