Added in prelimiary support for publications
There are still a couple of issues with this. The first is that it
doesn't actually display the items in the bibliography page. The second
is that the output from Zotero is not parsaable by the parser even
though it's perfectly standard BibTex.

The fix for that is to run `s/month = \(\w+\)/month = "\1"/g`

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>