Add support for Zotero BibTeX files
Zotero has a quirk where BibTeX files are written improperly and have
values like 'month = mar' where it should be 'month = {mar}'. Well, at
least that's what the javascript bibtex parser I'm using is
expecting. This transforms the input to ensure that it passes this
check.

Also added a little bit of documentation.

Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>