feat: updated for publication
After almost two years I finally decided to pull the trigger and public this
site. I had to disable the projects and CV page for a while as those aren't
ready. There also is a problem with my resume that I need to fix, but for the
most part it works.

Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>