fix: remove old content, fix mailcheck
This is a little bit more of the process of removing old content from the
server, namely an old bandwidth warning from when I used to serve the web page
from my cable modem in Pittsburgh about 10 years ago. Also some old pictures of
our house in Pittsburgh. This also applies a fix to the mailcheck code so it
doesn't get overwritten on every push.

In the long run if I'm moving to a docker based infrastructure I'll need to
find a better alternative for managing the data from my long running mail
tracking system.

Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>