fix: undo problems with shell explansion for mailcheck
The mailcheck code had a path that used shell expansion in a location where I
couldn't actually use shell expansion.

The correct solution is to make these apps take the path to the output/input
file from the command line, but that requires a tiny bit more effort that I'm
capable of at this time of night.

Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>