chore(cleanup): More migration and cleanup of files
There were a few directories that were permanently stored in public that
shouldn't have been there. In theory, the public directory should be entirely
rendered content.

There are still a few things in that directory that are in git, but it's
getting much better.

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>