refactor(weblog): clean up project code
More work on cleaning up the code. Mainly moving the weblog stuff into a
directory that indicates that it is a weblog. Also created a paginated front
page for the weblog. It appears that this may have introduced an issue with the
RSS feeds that I'll need to dig into.

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>