feat(resume): add shortcode for phone link
On my public resume I don't have my phone number because it's hard to
screen calls even with Google Voice. However, I wanted an easy way to
toggle this on and off for future use. This creates a couple of
shortcodes that I can use to add in infolinks on my resume, `infolink`
and `phonelink`. `phonelink` only renders when the site parameter
`resumeShowPhone` is set to `true`.

There's probably some fancy way that I could've tackled this with
partials or nested shortcodes (is that even possible), but this works
well enough for now and makes it easy for me to toggle it on and off.

It would be nice if there were a way that I could set `resumeShowPhone`
from the command line, but at the current time that does not seem
possible so I'll need to toggle the configuration when it needs to show
up.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>