fix(templates): Fix navbar and weblog listing rendering
This applies a couple of stop-gap fixes to the environment to help out with
rendering of the navbar, which had disappeared completely and also provide
something of an index of blog post articles without needing to list all of the
top 10 blog posts on the same page. In the medium-term I should probably roll
some of these changes into my actual theme, but this works to correct the minor
issues.

DCO 1.0 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>