fix(index): update current affiliation
This fixes the sidebar images of my current and past affiliations to be
reflective of the current state. It's getting long enough that in the future I
should consider just scrapping this section or finding an alternative way to
present the content. There still is boilerplate text that I need to refactor -
when that happens it will be refactored so it's not quite as role specific.

closes #9

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>