fix(home): Correct title on home page
This is a minor tweak that updates the title on the home page to have it
reflect the title that I use on my resume and LinkedIn. While I still delve
into software engineering and HCI, they're a smaller part of the role that I
take on now.

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>