fix(actions): ignore cache on GitHub Actions build
This is a very minor tweak that tells GitHub Actions to ignore the cache when
rebuilding the site. While this will result it in taking a little bit longer,
it hopeful should help fix some of the problems with content hanging around.

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>