fix(markdown): allow unsafe HTML
Hugo v0.60 changed the markdown renderer to Goldmark and in the process
caused a lot of things to break. Notably, my resume. This should be enough
of a workaround for my resume to render properly.

Also, this is, perhaps, a nice reminder, that my development version of
Hugo and the version used to render the website should be the same
version.

For more information about the change to Hugo see:
https://gohugo.io/news/0.60.0-relnotes/

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>