fix(actions): clear cache during site rebuild
This clears the cache of the project during any site rebuild. I think this
might help with some of the stale content problems that I'm seeing on Hugo. It
also removes the built files, which should make the rest of the process a bit
easier too, at the expense of a slower rebuild.

DCO 1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>