fix(comments): re-enable IntenseDebate
There was a bug that was introduced with some version of Hugo that
started lowercasing all of the keys in `Site.Params`, so a check like
`(isset .Site.Params "commentsEngine")` would fail because although that
might've been the casing in the configuration file, the keys were
converted to lowercase.

This should effectively re-enable IntenseDebate for comments on my
weblog. After four years of it being broken.

Closes #11

DCO-1.1-Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>