chore(emacs): disable lockfiles
There's some weird issues with Hugo and lockfiles from Emacs, which
meant that I had to disable lockfiles in this project. This was,
incidentally, the first time in 25 years of using Emacs that I had
heard of `.dir_locals.el`. It's great that I keep on learning new
things in Emacs.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>